Módulo base reestruturado
Módulo base reestruturado, contando agora com os submódulos 'ide' e
'item'. Dentro do 'item' há também uma nova subdivisão, o módulo
'imposto'.
