+++ nfe/src/base/ide/operacao.rs
