+++ nfe/src/base/mod.rs
@@ -9,14 +9,10 @@ use std::fs::File;
 use std::io::Read;
 use std::str::FromStr;
 pub mod dest;
-pub mod emissao;
 pub mod emit;
 pub mod endereco;
 pub mod ide;
-pub mod imposto;
 pub mod item;
-pub mod operacao;
-pub mod produto;
 pub mod versao;
 use dest::Destinatario;
 use emit::Emitente;
